Fix build error

+++ src/medium/maximum_product_subarray/solution.cpp
@@ -19,7 +19,7 @@ int Solution::maxProduct(vector<int>& nums) {
     int min_so_far = nums[0];
     int max_prod = max_so_far;
 
-    for (int i = 1; i < nums.size(); i++) {
+    for (size_t i = 1; i < nums.size(); i++) {
         int curr = nums[i];
         int temp_max = max(curr, max(max_so_far * curr, min_so_far * curr));
         min_so_far = min(curr, min(max_so_far * curr, min_so_far * curr));
